+++ server/storage/src/main/java/org/bithon/server/storage/datasource/input/IInputSource.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.source;
+package org.bithon.server.storage.datasource.input;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import org.bithon.server.storage.datasource.DataSourceSchema;
