+++ server/storage/src/main/java/org/bithon/server/storage/datasource/input/filter/IInputRowFilter.java
@@ -14,29 +14,23 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.transformer;
+package org.bithon.server.storage.datasource.input.filter;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import org.bithon.server.storage.datasource.input.IInputRow;
 
 /**
- * A transformer allows adding new fields to input rows.
- * Each one has a "name" (the name of the new field) which can be referred to by {@link org.bithon.server.storage.datasource.dimension.IDimensionSpec},
- * or {@link org.bithon.server.storage.datasource.aggregator.spec.IMetricSpec}.
- * <p>
- * The transformer produces values for this new field based on looking at the entire input row.
- *
- * @author Frank Chen
+ * @author frankchen
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
 @JsonSubTypes(value = {
-    @JsonSubTypes.Type(name = "mapping", value = MappingTransformer.class),
-    @JsonSubTypes.Type(name = "splitter", value = SplitterTransformer.class),
-    @JsonSubTypes.Type(name = "chain", value = ChainTransformer.class),
-    @JsonSubTypes.Type(name = "add", value = AddFieldTransformer.class)
+    @JsonSubTypes.Type(name = "endsWith", value = EndsWithFilter.class),
+    @JsonSubTypes.Type(name = "==", value = EqualFilter.class),
+    @JsonSubTypes.Type(name = ">", value = GreaterThanFilter.class),
+    @JsonSubTypes.Type(name = "or", value = OrFilter.class)
 })
-public interface ITransformer {
+public interface IInputRowFilter {
 
-    void transform(IInputRow inputRow);
+    boolean shouldInclude(IInputRow inputRow);
 }
