+++ server/storage/src/main/java/org/bithon/server/storage/datasource/input/flatten/TreePathFlattener.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.flatten;
+package org.bithon.server.storage.datasource.input.flatten;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
