+++ server/storage/src/main/java/org/bithon/server/storage/datasource/input/transformer/ITransformer.java
@@ -14,23 +14,29 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.filter;
+package org.bithon.server.storage.datasource.input.transformer;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import org.bithon.server.storage.datasource.input.IInputRow;
 
 /**
- * @author frankchen
+ * A transformer allows adding new fields to input rows.
+ * Each one has a "name" (the name of the new field) which can be referred to by {@link org.bithon.server.storage.datasource.dimension.IDimensionSpec},
+ * or {@link org.bithon.server.storage.datasource.aggregator.spec.IMetricSpec}.
+ * <p>
+ * The transformer produces values for this new field based on looking at the entire input row.
+ *
+ * @author Frank Chen
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
 @JsonSubTypes(value = {
-    @JsonSubTypes.Type(name = "endsWith", value = EndsWithFilter.class),
-    @JsonSubTypes.Type(name = "==", value = EqualFilter.class),
-    @JsonSubTypes.Type(name = ">", value = GreaterThanFilter.class),
-    @JsonSubTypes.Type(name = "or", value = OrFilter.class)
+    @JsonSubTypes.Type(name = "mapping", value = MappingTransformer.class),
+    @JsonSubTypes.Type(name = "splitter", value = SplitterTransformer.class),
+    @JsonSubTypes.Type(name = "chain", value = ChainTransformer.class),
+    @JsonSubTypes.Type(name = "add", value = AddFieldTransformer.class)
 })
-public interface IInputRowFilter {
+public interface ITransformer {
 
-    boolean shouldInclude(IInputRow inputRow);
+    void transform(IInputRow inputRow);
 }
