+++ server/storage/src/test/java/org/bithon/server/storage/datasource/transformer/ChainTransformerTest.java
@@ -20,6 +20,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableMap;
 import org.bithon.server.storage.datasource.input.InputRow;
+import org.bithon.server.storage.datasource.input.transformer.ChainTransformer;
+import org.bithon.server.storage.datasource.input.transformer.ITransformer;
+import org.bithon.server.storage.datasource.input.transformer.SplitterTransformer;
 import org.junit.Assert;
 import org.junit.Test;
